backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
python-constraint==1.4.0
psycopg2-binary==2.9.9
reportlab==4.1.0
openpyxl==3.1.3
django-cors-headers==4.3.0
Pillow==10.4.0
celery==5.5.3
psycopg2==2.9.9
